@@ -1,5 +1,5 @@
-# -*- encoding: utf-8 -*-
-$:.push File.expand_path('../lib', __FILE__)
+
+$LOAD_PATH.push File.expand_path('../lib', __FILE__)
 require 'casino/version'
 
 Gem::Specification.new do |s|
@@ -14,7 +14,7 @@ Gem::Specification.new do |s|
 
   s.files         = `git ls-files`.split("\n")
   s.test_files    = `git ls-files -- {test,spec,features}/*`.split("\n")
-  s.executables   = `git ls-files -- bin/*`.split("\n").map{ |f| File.basename(f) }
+  s.executables   = `git ls-files -- bin/*`.split("\n").map { |f| File.basename(f) }
   s.require_paths = ['lib']
 
   sign_file = File.expand_path '~/.gem/casino-private_key.pem'
@@ -23,25 +23,15 @@ Gem::Specification.new do |s|
     s.cert_chain  = ['casino-public_cert.pem']
   end
 
-  s.add_development_dependency 'capybara', '~> 2.1'
-  s.add_development_dependency 'rake', '~> 10.0'
-  s.add_development_dependency 'rspec', '~> 3.0'
-  s.add_development_dependency 'rspec-its', '~> 1.0'
-  s.add_development_dependency 'rspec-rails', '~> 3.0'
-  s.add_development_dependency 'sqlite3', '~> 1.3'
-  s.add_development_dependency 'factory_girl', '~> 4.1'
-  s.add_development_dependency 'webmock', '~> 1.9'
-  s.add_development_dependency 'coveralls', '~> 0.7'
-
-  s.add_runtime_dependency 'rails', '>= 4.1.0', '< 4.3.0'
-  s.add_runtime_dependency 'sass-rails', '>= 4.0.0', '< 6.0.0'
-  s.add_runtime_dependency 'addressable', '~> 2.3'
-  s.add_runtime_dependency 'terminal-table', '~> 1.4'
-  s.add_runtime_dependency 'useragent', '~> 0.4'
-  s.add_runtime_dependency 'faraday', '~> 0.8'
-  s.add_runtime_dependency 'rotp', '~> 2.0'
-  s.add_runtime_dependency 'grape', '~> 0.8'
-  s.add_runtime_dependency 'grape-entity', '~> 0.4'
-  s.add_runtime_dependency 'rqrcode_png', '~> 0.1'
+  s.add_runtime_dependency 'addressable', '>= 2.3'
+  s.add_runtime_dependency 'faraday', '>= 0.8'
+  s.add_runtime_dependency 'grape', '>= 0.8'
+  s.add_runtime_dependency 'grape-entity', '>= 0.4'
   s.add_runtime_dependency 'kaminari', '~> 0.16'
+  s.add_runtime_dependency 'rails', '>= 4.2'
+  s.add_runtime_dependency 'rotp', '>= 2.0'
+  s.add_runtime_dependency 'rqrcode_png', '>= 0.1'
+  s.add_runtime_dependency 'sass-rails', '>= 4.0.0'
+  s.add_runtime_dependency 'terminal-table', '>= 1.4'
+  s.add_runtime_dependency 'useragent', '>= 0.4'
 end
