@@ -1,15 +1,14 @@
-class CASino::AuthTokenTicket < ActiveRecord::Base
+class CASino::AuthTokenTicket < CASino::ApplicationRecord
   include CASino::ModelConcern::Ticket
   include CASino::ModelConcern::ConsumableTicket
 
   self.ticket_prefix = 'ATT'.freeze
 
   def self.cleanup
-    delete_all(['created_at < ?', CASino.config.auth_token_ticket[:lifetime].seconds.ago])
+    where(['created_at < ?', CASino.config.auth_token_ticket[:lifetime].seconds.ago]).delete_all
   end
 
   def expired?
-    (Time.now - (self.created_at || Time.now)) > CASino.config.auth_token_ticket[:lifetime].seconds
+    (Time.now - (created_at || Time.now)) > CASino.config.auth_token_ticket[:lifetime].seconds
   end
-
 end
