@@ -1,6 +1,6 @@
 require 'addressable/uri'
 
-class CASino::ServiceTicket < ActiveRecord::Base
+class CASino::ServiceTicket < CASino::ApplicationRecord
   include CASino::ModelConcern::Ticket
 
   self.ticket_prefix = 'ST'.freeze
@@ -10,15 +10,15 @@ class CASino::ServiceTicket < ActiveRecord::Base
   has_many :proxy_granting_tickets, as: :granter, dependent: :destroy
 
   def self.cleanup_unconsumed
-    self.delete_all(['created_at < ? AND consumed = ?', CASino.config.service_ticket[:lifetime_unconsumed].seconds.ago, false])
+    where(['created_at < ? AND consumed = ?', CASino.config.service_ticket[:lifetime_unconsumed].seconds.ago, false]).delete_all
   end
 
   def self.cleanup_consumed
-    self.destroy_all(['(ticket_granting_ticket_id IS NULL OR created_at < ?) AND consumed = ?', CASino.config.service_ticket[:lifetime_consumed].seconds.ago, true])
+    where(['(ticket_granting_ticket_id IS NULL OR created_at < ?) AND consumed = ?', CASino.config.service_ticket[:lifetime_consumed].seconds.ago, true]).destroy_all
   end
 
   def self.cleanup_consumed_hard
-    self.delete_all(['created_at < ? AND consumed = ?', (CASino.config.service_ticket[:lifetime_consumed] * 2).seconds.ago, true])
+    where(['created_at < ? AND consumed = ?', (CASino.config.service_ticket[:lifetime_consumed] * 2).seconds.ago, true]).delete_all
   end
 
   def service=(service)
@@ -27,21 +27,22 @@ class CASino::ServiceTicket < ActiveRecord::Base
   end
 
   def service_with_ticket_url
-    service_uri = Addressable::URI.parse(self.service)
-    service_uri.query_values = (service_uri.query_values(Array) || []) << ['ticket', self.ticket]
+    service_uri = Addressable::URI.parse(service)
+    service_uri.query_values = (service_uri.query_values(Array) || []) << ['ticket', ticket]
     service_uri.to_s
   end
 
   def expired?
     lifetime = if consumed?
-      CASino.config.service_ticket[:lifetime_consumed]
-    else
-      CASino.config.service_ticket[:lifetime_unconsumed]
-    end
-    (Time.now - (self.created_at || Time.now)) > lifetime
+                 CASino.config.service_ticket[:lifetime_consumed]
+               else
+                 CASino.config.service_ticket[:lifetime_unconsumed]
+               end
+    (Time.now - (created_at || Time.now)) > lifetime
   end
 
   private
+
   def send_single_sign_out_notification
     notifier = SingleSignOutNotifier.new(self)
     notifier.notify
