@@ -1,11 +1,11 @@
 # In order to support pre-2.0 installations of CASino that included CASinoCore,
 # we must rebuild the un-namespaced CASinoCore schema so that we can upgrade
-class CreateCoreSchema < ActiveRecord::Migration
+class CreateCoreSchema < ActiveRecord::Migration[4.1]
   CoreTables = %w{login_tickets proxy_granting_tickets proxy_tickets service_rules service_tickets ticket_granting_tickets two_factor_authenticators users}
 
   def up
     CoreTables.each do |table_name|
-      if !ActiveRecord::Base.connection.table_exists? table_name
+      unless connection.data_source_exists? table_name
         send "create_#{table_name}"
       end
     end
