@@ -1,4 +1,4 @@
-class ChangeServiceToText < ActiveRecord::Migration
+class ChangeServiceToText < ActiveRecord::Migration[4.1]
   def change
     change_column :casino_proxy_tickets, :service, :text
     change_column :casino_service_tickets, :service, :text
