@@ -1,4 +1,4 @@
-class FixLengthOfTextFields < ActiveRecord::Migration
+class FixLengthOfTextFields < ActiveRecord::Migration[4.1]
   def change
     change_column :casino_proxy_tickets, :service, :text, :limit => nil
     change_column :casino_service_tickets, :service, :text, :limit => nil
