@@ -5,16 +5,16 @@ describe CASino::LoginAttemptsController do
 
   describe 'GET #index' do
     context 'with ticket granting ticket' do
-      let(:ticket_granting_ticket) { FactoryGirl.create :ticket_granting_ticket }
-      let(:login_attempt) { FactoryGirl.create :login_attempt, user: ticket_granting_ticket.user }
+      let(:ticket_granting_ticket) { FactoryBot.create :ticket_granting_ticket }
+      let(:login_attempt) { FactoryBot.create :login_attempt, user: ticket_granting_ticket.user }
       let(:old_login_attempt) do
-        FactoryGirl.create :login_attempt, user: ticket_granting_ticket.user, created_at: 10.weeks.ago
+        FactoryBot.create :login_attempt, user: ticket_granting_ticket.user, created_at: 10.weeks.ago
       end
 
       before do
         sign_in(ticket_granting_ticket)
         login_attempt.touch
-        FactoryGirl.create :login_attempt
+        FactoryBot.create :login_attempt
       end
 
       it 'assigns current users login attempts @login_attempts' do
