@@ -1,7 +1,9 @@
 describe CASino::ServiceTicketsController do
+  routes { CASino::Engine.routes }
+
   describe 'GET "validate"' do
-    let(:request_options) { params.merge(use_route: :casino) }
-    let(:service_ticket) { FactoryGirl.create :service_ticket }
+    let(:request_options) { {params: params} }
+    let(:service_ticket) { FactoryBot.create :service_ticket }
     let(:service) { service_ticket.service }
     let(:parameters) { { service: service, ticket: service_ticket.ticket }}
     let(:params) { parameters }
@@ -14,12 +16,12 @@ describe CASino::ServiceTicketsController do
     context 'with an unconsumed service ticket' do
       context 'without renew flag' do
         it 'consumes the service ticket' do
-          get :validate, request_options
+          get :validate, **request_options
           service_ticket.reload.consumed.should == true
         end
 
         it 'answers with the expected response text' do
-          get :validate, request_options
+          get :validate, **request_options
           response.body.should == response_text_success
         end
       end
@@ -29,12 +31,12 @@ describe CASino::ServiceTicketsController do
 
         context 'with a service ticket without issued_from_credentials flag' do
           it 'consumes the service ticket' do
-            get :validate, request_options
+            get :validate, **request_options
             service_ticket.reload.consumed.should == true
           end
 
           it 'answers with the expected response text' do
-            get :validate, request_options
+            get :validate, **request_options
             response.body.should == response_text_failure
           end
         end
@@ -45,12 +47,12 @@ describe CASino::ServiceTicketsController do
           end
 
           it 'consumes the service ticket' do
-            get :validate, request_options
+            get :validate, **request_options
             service_ticket.reload.consumed.should == true
           end
 
           it 'answers with the expected response text' do
-            get :validate, request_options
+            get :validate, **request_options
             response.body.should == response_text_success
           end
         end
@@ -63,7 +65,7 @@ describe CASino::ServiceTicketsController do
       end
 
       it 'answers with the expected response text' do
-        get :validate, request_options
+        get :validate, **request_options
         response.body.should == response_text_failure
       end
     end
