@@ -5,8 +5,9 @@ describe CASino::SessionsController do
 
   routes { CASino::Engine.routes }
 
-  let(:params) { { } }
-  let(:user_agent) { 'YOLOBrowser 420.00'}
+  let(:params) { {} }
+  let(:request_options) { {params: params} }
+  let(:user_agent) { 'YOLOBrowser 420.00' }
 
   before(:each) do
     request.user_agent = user_agent
@@ -15,21 +16,21 @@ describe CASino::SessionsController do
   describe 'GET "new"' do
     context 'with a not allowed service' do
       before(:each) do
-        FactoryGirl.create :service_rule, :regex, url: '^https://.*'
+        FactoryBot.create :service_rule, :regex, url: '^https://.*'
       end
 
       let(:service) { 'http://www.example.org/' }
       let(:params) { { service: service } }
 
       it 'renders the service_not_allowed template' do
-        get :new, params
+        get :new, **request_options
         response.should render_template(:service_not_allowed)
       end
     end
 
     context 'when logged out' do
       it 'renders the new template' do
-        get :new, params
+        get :new, **request_options
         response.should render_template(:new)
       end
 
@@ -39,7 +40,7 @@ describe CASino::SessionsController do
           let(:params) { { service: service, gateway: 'true' } }
 
           it 'redirects to the service' do
-            get :new, params
+            get :new, **request_options
             response.should redirect_to(service)
           end
         end
@@ -48,7 +49,7 @@ describe CASino::SessionsController do
           let(:params) { { gateway: 'true' } }
 
           it 'renders the new template' do
-            get :new, params
+            get :new, **request_options
             response.should render_template(:new)
           end
         end
@@ -56,17 +57,17 @@ describe CASino::SessionsController do
     end
 
     context 'when logged in' do
-      let(:ticket_granting_ticket) { FactoryGirl.create :ticket_granting_ticket }
+      let(:ticket_granting_ticket) { FactoryBot.create :ticket_granting_ticket }
 
       before(:each) do
         sign_in(ticket_granting_ticket)
       end
 
       context 'when two-factor authentication is pending' do
-        let(:ticket_granting_ticket) { FactoryGirl.create :ticket_granting_ticket, :awaiting_two_factor_authentication }
+        let(:ticket_granting_ticket) { FactoryBot.create :ticket_granting_ticket, :awaiting_two_factor_authentication }
 
         it 'renders the new template' do
-          get :new, params
+          get :new, **request_options
           response.should render_template(:new)
         end
       end
@@ -78,7 +79,7 @@ describe CASino::SessionsController do
         end
 
         it 'renders the new template' do
-          get :new, params
+          get :new, **request_options
           response.should render_template(:new)
         end
       end
@@ -88,24 +89,24 @@ describe CASino::SessionsController do
         let(:params) { { service: service } }
 
         it 'redirects to the service' do
-          get :new, params
+          get :new, **request_options
           response.location.should =~ /^#{Regexp.escape service}\?ticket=ST-/
         end
 
         it 'generates a service ticket' do
-          lambda do
-            get :new, params
-          end.should change(CASino::ServiceTicket, :count).by(1)
+          -> { get :new, **request_options }.should change(CASino::ServiceTicket, :count).by(1)
         end
 
         it 'does not set the issued_from_credentials flag on the service ticket' do
-          get :new, params
+          get :new, **request_options
           CASino::ServiceTicket.last.should_not be_issued_from_credentials
         end
 
         context 'with renew parameter' do
+          let(:params) { super().merge(renew: 'true') }
+
           it 'renders the new template' do
-            get :new, params.merge(renew: 'true')
+            get :new, **request_options
             response.should render_template(:new)
           end
         end
@@ -116,7 +117,7 @@ describe CASino::SessionsController do
         let(:params) { { service: service } }
 
         it 'does not remove the attributes' do
-          get :new, params
+          get :new, **request_options
           response.location.should =~ /^#{Regexp.escape service}&ticket=ST-/
         end
       end
@@ -126,21 +127,19 @@ describe CASino::SessionsController do
         let(:params) { { service: service } }
 
         it 'redirects to the session overview' do
-          get :new, params
+          get :new, **request_options
           response.should redirect_to(sessions_path)
         end
       end
 
       context 'without a service' do
         it 'redirects to the session overview' do
-          get :new, params
+          get :new, **request_options
           response.should redirect_to(sessions_path)
         end
 
         it 'does not generate a service ticket' do
-          lambda do
-            get :new, params
-          end.should change(CASino::ServiceTicket, :count).by(0)
+          -> { get :new, **request_options }.should change(CASino::ServiceTicket, :count).by(0)
         end
 
         context 'with a changed browser' do
@@ -151,7 +150,7 @@ describe CASino::SessionsController do
           end
 
           it 'renders the new template' do
-            get :new, params
+            get :new, **request_options
             response.should render_template(:new)
           end
         end
@@ -160,7 +159,7 @@ describe CASino::SessionsController do
 
     context 'with an unsupported format' do
       it 'sets the status code to 406' do
-        get :new, use_route: :casino, format: :xml
+        get :new, format: :xml
         response.status.should == 406
       end
     end
@@ -169,41 +168,41 @@ describe CASino::SessionsController do
   describe 'POST "create"' do
     context 'without a valid login ticket' do
       it 'renders the new template' do
-        post :create, params
+        post :create, **request_options
         response.should render_template(:new)
       end
     end
 
     context 'with an expired login ticket' do
-      let(:expired_login_ticket) { FactoryGirl.create :login_ticket, :expired }
-      let(:params) { { lt: expired_login_ticket.ticket }}
+      let(:expired_login_ticket) { FactoryBot.create :login_ticket, :expired }
+      let(:params) { { lt: expired_login_ticket.ticket } }
 
       it 'renders the new template' do
-        post :create, params
+        post :create, **request_options
         response.should render_template(:new)
       end
     end
 
     context 'with a valid login ticket' do
-      let(:login_ticket) { FactoryGirl.create :login_ticket }
+      let(:login_ticket) { FactoryBot.create :login_ticket }
       let(:username) { 'testuser' }
-      let(:params) { { lt: login_ticket.ticket, username: username, password: 'wrrooonnng' }}
-      let!(:user) { FactoryGirl.create :user, username: username }
+      let(:params) { { lt: login_ticket.ticket, username: username, password: 'wrrooonnng' } }
+      let!(:user) { FactoryBot.create :user, username: username }
 
       context 'with invalid credentials' do
         it 'renders the new template' do
-          post :create, params
+          post :create, **request_options
           response.should render_template(:new)
         end
 
         it 'creates session log' do
           expect do
-            post :create, params
+            post :create, **request_options
           end.to change { CASino::LoginAttempt.count }.by 1
         end
 
         it 'assigns session log the correct attributes' do
-          post :create, params
+          post :create, **request_options
 
           expect(CASino::LoginAttempt.last.user).to eq user
           expect(CASino::LoginAttempt.last.successful).to eq false
@@ -217,24 +216,24 @@ describe CASino::SessionsController do
         let(:params) { { lt: login_ticket.ticket, username: username, password: 'foobar123', service: service } }
 
         it 'creates a cookie' do
-          post :create, params
+          post :create, **request_options
           response.cookies['tgt'].should_not be_nil
         end
 
         it 'saves user_ip' do
-          post :create, params
+          post :create, **request_options
           tgt = CASino::TicketGrantingTicket.last
           tgt.user_ip.should == '0.0.0.0'
         end
 
         it 'creates session log' do
           expect do
-            post :create, params
+            post :create, **request_options
           end.to change { CASino::LoginAttempt.count }.by 1
         end
 
         it 'assigns session log the correct attributes' do
-          post :create, params
+          post :create, **request_options
 
           expect(CASino::LoginAttempt.last.user.username).to eq username
           expect(CASino::LoginAttempt.last.successful).to eq true
@@ -249,12 +248,12 @@ describe CASino::SessionsController do
           end
 
           it 'creates a cookie with an expiration date set' do
-            post :create, params
+            post :create, **request_options
             cookie_jar['tgt']['expires'].should be_kind_of(Time)
           end
 
           it 'creates a long-term ticket-granting ticket' do
-            post :create, params
+            post :create, **request_options
             tgt = CASino::TicketGrantingTicket.last
             tgt.long_term.should == true
           end
@@ -262,22 +261,22 @@ describe CASino::SessionsController do
 
         context 'with two-factor authentication enabled' do
           let!(:user) { CASino::User.create! username: username, authenticator: authenticator }
-          let!(:two_factor_authenticator) { FactoryGirl.create :two_factor_authenticator, user: user }
+          let!(:two_factor_authenticator) { FactoryBot.create :two_factor_authenticator, user: user }
 
           it 'renders the validate_otp template' do
-            post :create, params
+            post :create, **request_options
             response.should render_template(:validate_otp)
           end
         end
 
         context 'with a not allowed service' do
           before(:each) do
-            FactoryGirl.create :service_rule, :regex, url: '^https://.*'
+            FactoryBot.create :service_rule, :regex, url: '^https://.*'
           end
           let(:service) { 'http://www.example.org/' }
 
           it 'renders the service_not_allowed template' do
-            post :create, params
+            post :create, **request_options
             response.should render_template(:service_not_allowed)
           end
         end
@@ -290,7 +289,7 @@ describe CASino::SessionsController do
           end
 
           it 'renders the new template' do
-            post :create, params
+            post :create, **request_options
             response.should render_template(:new)
           end
         end
@@ -299,25 +298,21 @@ describe CASino::SessionsController do
           let(:service) { nil }
 
           it 'redirects to the session overview' do
-            post :create, params
+            post :create, **request_options
             response.should redirect_to(sessions_path)
           end
 
           it 'generates a ticket-granting ticket' do
-            lambda do
-              post :create, params
-            end.should change(CASino::TicketGrantingTicket, :count).by(1)
+            -> { post :create, **request_options }.should change(CASino::TicketGrantingTicket, :count).by(1)
           end
 
           context 'when the user does not exist yet' do
             it 'generates exactly one user' do
-              lambda do
-                post :create, params
-              end.should change(CASino::User, :count).by(1)
+              -> { post :create, **request_options }.should change(CASino::User, :count).by(1)
             end
 
             it 'sets the users attributes' do
-              post :create, params
+              post :create, **request_options
               user = CASino::User.last
               user.username.should == username
               user.authenticator.should == authenticator
@@ -328,14 +323,12 @@ describe CASino::SessionsController do
             let!(:user) { CASino::User.create! username: username, authenticator: authenticator }
 
             it 'does not regenerate the user' do
-              lambda do
-                post :create, params
-              end.should_not change(CASino::User, :count)
+              -> { post :create, **request_options }.should_not change(CASino::User, :count)
             end
 
             it 'updates the extra attributes' do
               lambda do
-                post :create, params
+                post :create, **request_options
                 user.reload
               end.should change(user, :extra_attributes)
             end
@@ -346,25 +339,21 @@ describe CASino::SessionsController do
           let(:service) { 'https://www.example.com' }
 
           it 'redirects to the service' do
-            post :create, params
+            post :create, **request_options
             response.location.should =~ /^#{Regexp.escape service}\/\?ticket=ST-/
           end
 
           it 'generates a service ticket' do
-            lambda do
-              post :create, params
-            end.should change(CASino::ServiceTicket, :count).by(1)
+            -> { post :create, **request_options }.should change(CASino::ServiceTicket, :count).by(1)
           end
 
           it 'does set the issued_from_credentials flag on the service ticket' do
-            post :create, params
+            post :create, **request_options
             CASino::ServiceTicket.last.should be_issued_from_credentials
           end
 
           it 'generates a ticket-granting ticket' do
-            lambda do
-              post :create, params
-            end.should change(CASino::TicketGrantingTicket, :count).by(1)
+            -> { post :create, **request_options }.should change(CASino::TicketGrantingTicket, :count).by(1)
           end
         end
       end
@@ -373,16 +362,16 @@ describe CASino::SessionsController do
 
   describe 'POST "validate_otp"' do
     context 'with an existing ticket-granting ticket' do
-      let(:ticket_granting_ticket) { FactoryGirl.create :ticket_granting_ticket, :awaiting_two_factor_authentication }
+      let(:ticket_granting_ticket) { FactoryBot.create :ticket_granting_ticket, :awaiting_two_factor_authentication }
       let(:user) { ticket_granting_ticket.user }
       let(:tgt) { ticket_granting_ticket.ticket }
       let(:user_agent) { ticket_granting_ticket.user_agent }
       let(:otp) { '123456' }
       let(:service) { 'http://www.example.com/testing' }
-      let(:params) { { tgt: tgt, otp: otp, service: service }}
+      let(:params) { { tgt: tgt, otp: otp, service: service } }
 
       context 'with an active authenticator' do
-        let!(:two_factor_authenticator) { FactoryGirl.create :two_factor_authenticator, user: user }
+        let!(:two_factor_authenticator) { FactoryBot.create :two_factor_authenticator, user: user }
 
         context 'with a valid OTP' do
           before(:each) do
@@ -390,12 +379,12 @@ describe CASino::SessionsController do
           end
 
           it 'redirects to the service' do
-            post :validate_otp, params
+            post :validate_otp, **request_options
             response.location.should =~ /^#{Regexp.escape service}\?ticket=ST-/
           end
 
           it 'does activate the ticket-granting ticket' do
-            post :validate_otp, params
+            post :validate_otp, **request_options
             ticket_granting_ticket.reload.should_not be_awaiting_two_factor_authentication
           end
 
@@ -408,19 +397,19 @@ describe CASino::SessionsController do
             end
 
             it 'creates a cookie with an expiration date set' do
-              post :validate_otp, params
+              post :validate_otp, **request_options
               cookie_jar['tgt']['expires'].should be_kind_of(Time)
             end
           end
 
           context 'with a not allowed service' do
             before(:each) do
-              FactoryGirl.create :service_rule, :regex, url: '^https://.*'
+              FactoryBot.create :service_rule, :regex, url: '^https://.*'
             end
             let(:service) { 'http://www.example.org/' }
 
             it 'renders the service_not_allowed template' do
-              post :validate_otp, params
+              post :validate_otp, **request_options
               response.should render_template(:service_not_allowed)
             end
           end
@@ -432,12 +421,12 @@ describe CASino::SessionsController do
           end
 
           it 'renders the validate_otp template' do
-            post :validate_otp, params
+            post :validate_otp, **request_options
             response.should render_template(:validate_otp)
           end
 
           it 'does not activate the ticket-granting ticket' do
-            post :validate_otp, params
+            post :validate_otp, **request_options
             ticket_granting_ticket.reload.should be_awaiting_two_factor_authentication
           end
         end
@@ -446,7 +435,7 @@ describe CASino::SessionsController do
 
     context 'without a ticket-granting ticket' do
       it 'redirects to the login page' do
-        post :validate_otp, params
+        post :validate_otp, **request_options
         response.should redirect_to(login_path)
       end
     end
@@ -454,22 +443,22 @@ describe CASino::SessionsController do
 
   describe 'GET "logout"' do
     let(:url) { nil }
-    let(:params) { { :url => url } }
+    let(:params) { { url: url } }
 
     context 'with an existing ticket-granting ticket' do
-      let(:ticket_granting_ticket) { FactoryGirl.create :ticket_granting_ticket }
+      let(:ticket_granting_ticket) { FactoryBot.create :ticket_granting_ticket }
 
       before(:each) do
         sign_in(ticket_granting_ticket)
       end
 
       it 'deletes the ticket-granting ticket' do
-        get :logout, params
-        CASino::TicketGrantingTicket.where(id: ticket_granting_ticket.id).first.should == nil
+        get :logout, **request_options
+        CASino::TicketGrantingTicket.where(id: ticket_granting_ticket.id).first.should.nil?
       end
 
       it 'renders the logout template' do
-        get :logout, params
+        get :logout, **request_options
         response.should render_template(:logout)
       end
 
@@ -477,34 +466,34 @@ describe CASino::SessionsController do
         let(:url) { 'http://www.example.com' }
 
         it 'assigns the URL' do
-          get :logout, params
+          get :logout, **request_options
           assigns(:url).should == url
         end
       end
 
       context 'with a service' do
-        let(:params) { { :service => url } }
+        let(:params) { { service: url } }
         let(:url) { 'http://www.example.org' }
 
         context 'when whitelisted' do
           it 'redirects to the service' do
-            get :logout, params
+            get :logout, **request_options
             response.should redirect_to(url)
           end
         end
 
         context 'when not whitelisted' do
           before(:each) do
-            FactoryGirl.create :service_rule, :regex, url: '^https://.*'
+            FactoryBot.create :service_rule, :regex, url: '^https://.*'
           end
 
           it 'renders the logout template' do
-            get :logout, params
+            get :logout, **request_options
             response.should render_template(:logout)
           end
 
           it 'does not assign the URL' do
-            get :logout, params
+            get :logout, **request_options
             assigns(:url).should be_nil
           end
         end
@@ -513,7 +502,7 @@ describe CASino::SessionsController do
 
     context 'without a ticket-granting ticket' do
       it 'renders the logout template' do
-        get :logout, params
+        get :logout, **request_options
         response.should render_template(:logout)
       end
     end
@@ -526,67 +515,67 @@ describe CASino::SessionsController do
       end
 
       describe 'two-factor authenticator settings' do
-        let(:ticket_granting_ticket) { FactoryGirl.create :ticket_granting_ticket }
+        let(:ticket_granting_ticket) { FactoryBot.create :ticket_granting_ticket }
         let(:user) { ticket_granting_ticket.user }
 
         context 'without a two-factor authenticator registered' do
           it 'does not assign any two-factor authenticators' do
-            get :index, params
+            get :index, **request_options
             assigns(:two_factor_authenticators).should == []
           end
         end
 
         context 'with an inactive two-factor authenticator' do
-          let!(:two_factor_authenticator) { FactoryGirl.create :two_factor_authenticator, :inactive, user: user }
+          let!(:two_factor_authenticator) { FactoryBot.create :two_factor_authenticator, :inactive, user: user }
 
           it 'does not assign any two-factor authenticators' do
-            get :index, params
+            get :index, **request_options
             assigns(:two_factor_authenticators).should == []
           end
         end
 
         context 'with a two-factor authenticator registered' do
-          let(:two_factor_authenticator) { FactoryGirl.create :two_factor_authenticator, user: user }
-          let!(:other_two_factor_authenticator) { FactoryGirl.create :two_factor_authenticator }
+          let(:two_factor_authenticator) { FactoryBot.create :two_factor_authenticator, user: user }
+          let!(:other_two_factor_authenticator) { FactoryBot.create :two_factor_authenticator }
 
           it 'does assign the two-factor authenticator' do
-            get :index, params
+            get :index, **request_options
             assigns(:two_factor_authenticators).should == [two_factor_authenticator]
           end
         end
       end
 
       describe 'sessions overview' do
-        let!(:other_ticket_granting_ticket) { FactoryGirl.create :ticket_granting_ticket }
+        let!(:other_ticket_granting_ticket) { FactoryBot.create :ticket_granting_ticket }
         let(:user) { other_ticket_granting_ticket.user }
 
         context 'as user owning the other ticket granting ticket' do
-          let(:ticket_granting_ticket) { FactoryGirl.create :ticket_granting_ticket, user: user }
+          let(:ticket_granting_ticket) { FactoryBot.create :ticket_granting_ticket, user: user }
 
           it 'assigns both ticket granting tickets' do
-            get :index, params
+            get :index, **request_options
             assigns(:ticket_granting_tickets).should == [ticket_granting_ticket, other_ticket_granting_ticket]
           end
         end
 
         context 'with a ticket-granting ticket with same username but different authenticator' do
-          let(:ticket_granting_ticket) { FactoryGirl.create :ticket_granting_ticket }
+          let(:ticket_granting_ticket) { FactoryBot.create :ticket_granting_ticket }
           let(:tgt) { ticket_granting_ticket.ticket }
 
           it 'does not assign the other ticket granting ticket' do
-            get :index, params
+            get :index, **request_options
             assigns(:ticket_granting_tickets).should == [ticket_granting_ticket]
           end
         end
       end
 
       describe 'last login attempts' do
-        let(:ticket_granting_ticket) { FactoryGirl.create :ticket_granting_ticket }
+        let(:ticket_granting_ticket) { FactoryBot.create :ticket_granting_ticket }
 
         let(:login_attempts) do
           6.times.map do |counter|
-            FactoryGirl.create :login_attempt, user: ticket_granting_ticket.user,
-                                               created_at: counter.minutes.ago
+            FactoryBot.create :login_attempt, user: ticket_granting_ticket.user,
+                                              created_at: counter.minutes.ago
           end
         end
 
@@ -597,7 +586,7 @@ describe CASino::SessionsController do
         end
 
         it 'assigns last five login attempts' do
-          get :index, params
+          get :index, **request_options
 
           expect(assigns(:login_attempts)).to eq login_attempts.sort_by(&:created_at).from(1).to(6).reverse
         end
@@ -606,14 +595,14 @@ describe CASino::SessionsController do
 
     context 'without a ticket-granting ticket' do
       it 'redirects to the login page' do
-        get :index, params
+        get :index, **request_options
         response.should redirect_to(login_path)
       end
     end
   end
 
   describe 'DELETE "destroy"' do
-    let(:owner_ticket_granting_ticket) { FactoryGirl.create :ticket_granting_ticket }
+    let(:owner_ticket_granting_ticket) { FactoryBot.create :ticket_granting_ticket }
     let(:user) { owner_ticket_granting_ticket.user }
 
     before(:each) do
@@ -621,54 +610,48 @@ describe CASino::SessionsController do
     end
 
     context 'with an existing ticket-granting ticket' do
-      let!(:ticket_granting_ticket) { FactoryGirl.create :ticket_granting_ticket, user: user }
-      let(:service_ticket) { FactoryGirl.create :service_ticket, ticket_granting_ticket: ticket_granting_ticket }
-      let(:consumed_service_ticket) { FactoryGirl.create :service_ticket, :consumed, ticket_granting_ticket: ticket_granting_ticket }
+      let!(:ticket_granting_ticket) { FactoryBot.create :ticket_granting_ticket, user: user }
+      let(:service_ticket) { FactoryBot.create :service_ticket, ticket_granting_ticket: ticket_granting_ticket }
+      let(:consumed_service_ticket) { FactoryBot.create :service_ticket, :consumed, ticket_granting_ticket: ticket_granting_ticket }
       let(:params) { { id: ticket_granting_ticket.id } }
 
       it 'deletes exactly one ticket-granting ticket' do
-        lambda do
-          delete :destroy, params
-        end.should change(CASino::TicketGrantingTicket, :count).by(-1)
+        -> { delete :destroy, **request_options }.should change(CASino::TicketGrantingTicket, :count).by(-1)
       end
 
       it 'deletes the ticket-granting ticket' do
-        delete :destroy, params
+        delete :destroy, **request_options
         CASino::TicketGrantingTicket.where(id: params[:id]).length.should == 0
       end
 
       it 'redirects to the session overview' do
-        delete :destroy, params
+        delete :destroy, **request_options
         response.should redirect_to(sessions_path)
       end
     end
 
     context 'with an invalid ticket-granting ticket' do
-      let(:params) { { id: 99999 } }
+      let(:params) { { id: 99_999 } }
       it 'does not delete a ticket-granting ticket' do
-        lambda do
-          delete :destroy, params
-        end.should_not change(CASino::TicketGrantingTicket, :count)
+        -> { delete :destroy, **request_options }.should_not change(CASino::TicketGrantingTicket, :count)
       end
 
       it 'redirects to the session overview' do
-        delete :destroy, params
+        delete :destroy, **request_options
         response.should redirect_to(sessions_path)
       end
     end
 
     context 'when trying to delete ticket-granting ticket of another user' do
-      let!(:ticket_granting_ticket) { FactoryGirl.create :ticket_granting_ticket }
+      let!(:ticket_granting_ticket) { FactoryBot.create :ticket_granting_ticket }
       let(:params) { { id: ticket_granting_ticket.id } }
 
       it 'does not delete a ticket-granting ticket' do
-        lambda do
-          delete :destroy, params
-        end.should_not change(CASino::TicketGrantingTicket, :count)
+        -> { delete :destroy, **request_options }.should_not change(CASino::TicketGrantingTicket, :count)
       end
 
       it 'redirects to the session overview' do
-        delete :destroy, params
+        delete :destroy, **request_options
         response.should redirect_to(sessions_path)
       end
     end
@@ -676,26 +659,24 @@ describe CASino::SessionsController do
 
   describe 'GET "destroy_others"' do
     let(:url) { nil }
-    let(:params) { { :service => url } }
+    let(:params) { { service: url } }
 
     context 'with an existing ticket-granting ticket' do
-      let(:user) { FactoryGirl.create :user }
-      let!(:other_users_ticket_granting_tickets) { FactoryGirl.create_list :ticket_granting_ticket, 3 }
-      let!(:other_ticket_granting_tickets) { FactoryGirl.create_list :ticket_granting_ticket, 3, user: user }
-      let!(:ticket_granting_ticket) { FactoryGirl.create :ticket_granting_ticket, user: user }
+      let(:user) { FactoryBot.create :user }
+      let!(:other_users_ticket_granting_tickets) { FactoryBot.create_list :ticket_granting_ticket, 3 }
+      let!(:other_ticket_granting_tickets) { FactoryBot.create_list :ticket_granting_ticket, 3, user: user }
+      let!(:ticket_granting_ticket) { FactoryBot.create :ticket_granting_ticket, user: user }
 
       before(:each) do
         sign_in(ticket_granting_ticket)
       end
 
       it 'deletes all other ticket-granting tickets' do
-        lambda do
-          get :destroy_others, params
-        end.should change(CASino::TicketGrantingTicket, :count).by(-3)
+        -> { get :destroy_others, **request_options }.should change(CASino::TicketGrantingTicket, :count).by(-3)
       end
 
       it 'redirects to the session overview' do
-        get :destroy_others, params
+        get :destroy_others, **request_options
         response.should redirect_to(sessions_path)
       end
 
@@ -703,7 +684,7 @@ describe CASino::SessionsController do
         let(:url) { 'http://www.example.com' }
 
         it 'redirects to the service' do
-          get :destroy_others, params
+          get :destroy_others, **request_options
           response.should redirect_to(url)
         end
       end
@@ -714,7 +695,7 @@ describe CASino::SessionsController do
         let(:url) { 'http://www.example.com' }
 
         it 'redirects to the service' do
-          get :destroy_others, params
+          get :destroy_others, **request_options
           response.should redirect_to(url)
         end
       end
