@@ -26,7 +26,7 @@ describe CASino::ServiceRule do
 
     context 'with a regex rule' do
       before(:each) do
-        FactoryGirl.create :service_rule, :regex, url: '^https://.*'
+        FactoryBot.create :service_rule, :regex, url: '^https://.*'
       end
 
       ['https://www.example.org/', 'https://www.google.com/'].each do |service_url|
@@ -45,7 +45,7 @@ describe CASino::ServiceRule do
     context 'with many regex rules' do
       before(:each) do
         100.times do |counter|
-          FactoryGirl.create :service_rule, :regex, url: "^https://www#{counter}.example.com"
+          FactoryBot.create :service_rule, :regex, url: "^https://www#{counter}.example.com"
         end
       end
 
@@ -60,7 +60,7 @@ describe CASino::ServiceRule do
 
     context 'with a non-regex rule' do
       before(:each) do
-        FactoryGirl.create :service_rule, url: 'https://www.google.com/foo'
+        FactoryBot.create :service_rule, url: 'https://www.google.com/foo'
       end
 
       ['https://www.google.com/foo'].each do |service_url|
