@@ -3,7 +3,7 @@ require 'spec_helper'
 describe CASino::TwoFactorAuthenticator do
   describe '.cleanup' do
     it 'deletes expired inactive two-factor authenticators' do
-      authenticator = FactoryGirl.create :two_factor_authenticator, :inactive
+      authenticator = FactoryBot.create :two_factor_authenticator, :inactive
       authenticator.created_at = 10.hours.ago
       authenticator.save!
       lambda do
@@ -12,7 +12,7 @@ describe CASino::TwoFactorAuthenticator do
     end
 
     it 'does not delete not expired inactive two-factor authenticators' do
-      authenticator = FactoryGirl.create :two_factor_authenticator, :inactive
+      authenticator = FactoryBot.create :two_factor_authenticator, :inactive
       authenticator.created_at = (CASino.config.two_factor_authenticator[:lifetime_inactive].seconds - 5).ago
       lambda do
         described_class.cleanup
@@ -20,7 +20,7 @@ describe CASino::TwoFactorAuthenticator do
     end
 
     it 'does not delete active two-factor authenticators' do
-      authenticator = FactoryGirl.create :two_factor_authenticator
+      authenticator = FactoryBot.create :two_factor_authenticator
       authenticator.created_at = 10.hours.ago
       authenticator.save!
       lambda do
