@@ -1,6 +1,6 @@
 require 'factory_girl'
 
-FactoryGirl.define do
+FactoryBot.define do
   factory :proxy_ticket, class: CASino::ProxyTicket do
     proxy_granting_ticket
     sequence :ticket do |n|
