@@ -1,6 +1,6 @@
 require 'factory_girl'
 
-FactoryGirl.define do
+FactoryBot.define do
   factory :service_rule, class: CASino::ServiceRule do
     sequence :order do |n|
       n
