@@ -23,6 +23,10 @@ module CASino::SessionsHelper
     tgt.user
   end
 
+  def current_authenticator_context
+    CASino.config.authenticator_context_builder.call(params, request)
+  end
+
   def ensure_signed_in
     redirect_to login_path unless signed_in?
   end
@@ -83,12 +87,12 @@ module CASino::SessionsHelper
   end
 
   def handle_signed_in_with_service(tgt, options)
-    if !service_allowed?(params[:service])
-      @service = params[:service]
-      render 'casino/sessions/service_not_allowed', status: 403
-    else
+    if service_allowed?(params[:service])
       url = acquire_service_ticket(tgt, params[:service], options).service_with_ticket_url
       redirect_to url, status: :see_other
+    else
+      @service = params[:service]
+      render 'casino/sessions/service_not_allowed', status: 403
     end
   end
 end
