@@ -3,11 +3,17 @@ require 'casino/authenticator'
 module CASino::AuthenticationProcessor
   extend ActiveSupport::Concern
 
-  def validate_login_credentials(username, password)
+  def validate_login_credentials(username, password, context = nil)
     authentication_result = nil
     authenticators.each do |authenticator_name, authenticator|
       begin
-        data = authenticator.validate(username, password)
+        credentials = [ username, password, context ]
+
+        # Old authenticators that don't accept a 3rd context parameter will have a validate
+        # method that only accepts 2 arguments, so check for that.
+        credentials.pop if authenticator.class.instance_method(:validate).arity == 2
+
+        data = authenticator.validate(*credentials)
       rescue CASino::Authenticator::AuthenticatorError => e
         Rails.logger.error "Authenticator '#{authenticator_name}' (#{authenticator.class}) raised an error: #{e}"
       end
