@@ -20,18 +20,18 @@ module CASino
     def install_migrations
       return unless options['migration']
 
-      rake 'casino:install:migrations'
+      generate 'casino:migration'
     end
 
     def copy_config_files
       return unless options['config_files']
 
       copy_file 'cas.yml', 'config/cas.yml'
-      copy_file 'casino_and_overrides.scss', 'app/assets/stylesheets/casino_and_overrides.scss'
+      copy_file 'casino_and_overrides.scss', "app/assets/stylesheets/#{namespace_name}/casino_and_overrides.scss".squeeze('/')
     end
 
     def insert_assets_loader
-      insert_into_file 'app/assets/javascripts/application.js', :after => %r{//= require +['"]?jquery_ujs['"]?} do
+      insert_into_file "app/assets/javascripts/#{namespace_name}/application.js".squeeze('/'), :after => %r{//= require +['"]?jquery_ujs['"]?} do
         "\n//= require casino"
       end
     end
@@ -43,5 +43,11 @@ module CASino
     def show_readme
       readme 'README'
     end
+
+    private
+      def namespace_name
+        Rails::Generators.namespace.to_s.underscore
+      end
+
   end
 end
